Add read only and write only option to load test.

+++ src/yb/benchmarks/yb_load_test_tool.cc
@@ -43,7 +43,7 @@ DEFINE_int64(
   load_test_num_rows, 50000,
   "Number of rows to insert");
 
-DEFINE_int64(
+DEFINE_int32(
   load_test_num_writer_threads, 4,
   "Number of writer threads");
 
@@ -69,7 +69,19 @@ DEFINE_int32(
 
 DEFINE_int32(
   load_test_num_tablets, 16,
-  "Number of tables to create in the table");
+  "Number of tablets to create in the table");
+
+DEFINE_bool(
+    load_test_reads_only, false,
+    "Only read the existing rows from the table.");
+
+DEFINE_bool(
+    load_test_writes_only, false,
+    "Writes a new set of rows into an existing table.");
+
+DEFINE_bool(
+    create_table, true,
+    "Should the table be created, its made false when either reads_only/writes_only is true.");
 
 using strings::Substitute;
 using std::atomic_long;
@@ -174,6 +186,7 @@ public:
   MultiThreadedAction(
     const string& description,
     int64 num_keys,
+    int64 start_key,
     int num_action_threads,
     int num_extra_threads,
     YBClient* client,
@@ -201,7 +214,8 @@ protected:
   virtual void RunStatsThread() = 0;
 
   string description_;
-  const int64 num_keys_;
+  const int64 num_keys_; // Total number of keys in the table after successful end of this action
+  const int64 start_key_; // Applies to write action for first insertion key index
   const int num_action_threads_;
   YBClient* client_;
   YBTable* table_;
@@ -215,13 +229,14 @@ protected:
 MultiThreadedAction::MultiThreadedAction(
     const string& description,
     int64 num_keys,
+    int64 start_key,
     int num_action_threads,
     int num_extra_threads,
     YBClient* client,
     YBTable* table,
     atomic_bool* stop_flag)
   : description_(description),
-    num_keys_(num_keys),
+    num_keys_(num_keys), start_key_(start_key),
     num_action_threads_(num_action_threads),
     client_(client),
     table_(table),
@@ -254,7 +269,7 @@ class MultiThreadedWriter : public MultiThreadedAction {
 // client and table are managed by the caller and their lifetime should be a superset of this
   // object's lifetime.
   MultiThreadedWriter(
-    int64 num_keys,
+    int64 num_keys, int64 start_key,
     int num_writer_threads,
     YBClient* client,
     YBTable* table,
@@ -280,13 +295,14 @@ class MultiThreadedWriter : public MultiThreadedAction {
 };
 
 MultiThreadedWriter::MultiThreadedWriter(
-    int64 num_keys,
+    int64 num_keys, int64 start_key,
     int num_writer_threads,
     YBClient* client,
     YBTable* table,
     atomic_bool* stop_flag)
-  : MultiThreadedAction("writers", num_keys, num_writer_threads, 2, client, table, stop_flag),
-    next_key_(0),
+  : MultiThreadedAction("writers", num_keys, start_key, num_writer_threads, 2, client, table,
+                          stop_flag),
+    next_key_(start_key),
     inserted_up_to_inclusive_(-1) {
 }
 
@@ -351,7 +367,7 @@ void MultiThreadedWriter::RunStatsThread() {
     running_threads_latch_.WaitFor(MonoDelta::FromSeconds(1));
     int64 current_key = next_key_.load();
     MicrosecondsInt64 current_time = GetMonoTimeMicros();
-    LOG(INFO) << "Wrote " << current_key << " rows ("
+    LOG(INFO) << "Wrote " << current_key - start_key_ << " rows ("
       << current_key * 1000000.0 / (current_time - start_time) << " writes/sec)"
       << ", contiguous insertion point: " << inserted_up_to_inclusive_.load()
       << ", write errors: " << failed_keys_.NumElements();
@@ -375,9 +391,40 @@ void MultiThreadedWriter::RunInsertionTrackerThread() {
 }
 
 // ------------------------------------------------------------------------------------------------
+// SingleThreadedScanner
+// ------------------------------------------------------------------------------------------------
+// TODO: Make MultiThreaded.
+class SingleThreadedScanner {
+public:
+  SingleThreadedScanner(YBTable* table) {
+    table_ = table;
+    num_rows_ = 0;
+  }
+
+  uint64 countRows();
+
+private:
+  uint64   num_rows_;
+  YBTable* table_;
+};
+
+uint64 SingleThreadedScanner::countRows() {
+  YBScanner scanner(table_);
+  CHECK_OK(scanner.Open());
+  vector<YBRowResult> results;
+  while (scanner.HasMoreRows()) {
+    CHECK_OK(scanner.NextBatch(&results));
+    num_rows_ += results.size();
+    results.clear();
+  }
+
+  LOG(INFO) << " num read rows = " << num_rows_;
+  return num_rows_;
+}
+
+// ------------------------------------------------------------------------------------------------
 // MultiThreadedReader
 // ------------------------------------------------------------------------------------------------
-
 class MultiThreadedReader : public MultiThreadedAction {
  public:
   MultiThreadedReader(
@@ -412,7 +459,7 @@ MultiThreadedReader::MultiThreadedReader(
     const KeyIndexSet* inserted_keys,
     const KeyIndexSet* failed_keys,
     atomic_bool* stop_flag)
-  : MultiThreadedAction("readers", num_keys, num_reader_threads, 1, client, table, stop_flag),
+  : MultiThreadedAction("readers", num_keys, 0, num_reader_threads, 1, client, table, stop_flag),
     insertion_point_(insertion_point),
     inserted_keys_(inserted_keys),
     failed_keys_(failed_keys),
@@ -542,7 +589,8 @@ int main(int argc, char* argv[]) {
   yb::ParseCommandLineFlags(&argc, &argv, true);
   yb::InitGoogleLoggingSafe(argv[0]);
 
-  LOG(INFO) << "num_keys = " << FLAGS_load_test_num_rows;
+  if (!FLAGS_load_test_reads_only)
+    LOG(INFO) << "num_keys = " << FLAGS_load_test_num_rows;
 
   for (int i = 0; i < FLAGS_load_test_num_iter; ++i) {
     shared_ptr<YBClient> client;
@@ -553,49 +601,63 @@ int main(int argc, char* argv[]) {
 
     const string table_name(FLAGS_load_test_table_name);
 
+    if (FLAGS_load_test_reads_only || FLAGS_load_test_writes_only) {
+      FLAGS_create_table = false;
+    }
+
     LOG(INFO) << "Checking if table '" << table_name << "' already exists";
     {
       YBSchema existing_schema;
       if (client->GetTableSchema(table_name, &existing_schema).ok()) {
-        LOG(INFO) << "Table '" << table_name << "' already exists, deleting";
-        // Table with the same name already exists, drop it.
-        CHECK_OK(client->DeleteTable(table_name));
+        if (FLAGS_create_table) {
+          LOG(INFO) << "Table '" << table_name << "' already exists, deleting";
+          // Table with the same name already exists, drop it.
+          CHECK_OK(client->DeleteTable(table_name));
+        }
       } else {
         LOG(INFO) << "Table '" << table_name << "' does not exist yet";
+
+        if (!FLAGS_create_table) {
+          LOG(ERROR) << "Exiting as the table was not asked to be created.";
+          return 0;
+        }
       }
     }
 
-    LOG(INFO) << "Building schema";
-    YBSchemaBuilder schemaBuilder;
-    schemaBuilder.AddColumn("k")->PrimaryKey()->Type(YBColumnSchema::STRING)->NotNull();
-    schemaBuilder.AddColumn("v")->Type(YBColumnSchema::STRING)->NotNull();
-    YBSchema schema;
-    CHECK_OK(schemaBuilder.Build(&schema));
+    if (FLAGS_create_table) {
+      LOG(INFO) << "Building schema";
+      YBSchemaBuilder schemaBuilder;
+      schemaBuilder.AddColumn("k")->PrimaryKey()->Type(YBColumnSchema::STRING)->NotNull();
+      schemaBuilder.AddColumn("v")->Type(YBColumnSchema::STRING)->NotNull();
+      YBSchema schema;
+      CHECK_OK(schemaBuilder.Build(&schema));
 
-    // Create the number of partitions based on the split keys.
-    vector<const YBPartialRow*> splits;
-    for (uint64_t i = 1; i < FLAGS_load_test_num_tablets; i++) {
-      YBPartialRow* row = schema.NewRow();
-      // We divide the interval between 0 and 2**64 into the requested number of intervals.
-      string split_key = FormatHex(
-        ((uint64_t) 1 << 62) * 4.0 * i / FLAGS_load_test_num_tablets);
-      LOG(INFO) << "split_key #" << i << "=" << split_key;
-      CHECK_OK(row->SetStringCopy(0, split_key));
-      splits.push_back(row);
-    }
+      // Create the number of partitions based on the split keys.
+      vector<const YBPartialRow *> splits;
+      for (uint64_t i = 1; i < FLAGS_load_test_num_tablets; i++) {
+        YBPartialRow *row = schema.NewRow();
+        // We divide the interval between 0 and 2**64 into the requested number of intervals.
+        string split_key = FormatHex(
+            ((uint64_t) 1 << 62) * 4.0 * i / (FLAGS_load_test_num_tablets));
+        LOG(INFO) << "split_key #" << i << "=" << split_key;
+        CHECK_OK(row->SetStringCopy(0, split_key));
+        splits.push_back(row);
+      }
 
-    LOG(INFO) << "Creating table";
-    gscoped_ptr<YBTableCreator> table_creator(client->NewTableCreator());
-    Status table_creation_status = table_creator->table_name(table_name).schema(&schema)
-      .split_rows(splits)
-      .num_replicas(FLAGS_load_test_table_num_replicas).Create();
-    if (!table_creation_status.ok()) {
-      LOG(INFO) << "Table creation status message: " << table_creation_status.message().ToString();
-    }
+      LOG(INFO) << "Creating table";
+      gscoped_ptr<YBTableCreator> table_creator(client->NewTableCreator());
+      Status table_creation_status = table_creator->table_name(table_name).schema(&schema)
+          .split_rows(splits)
+          .num_replicas(FLAGS_load_test_table_num_replicas).Create();
+      if (!table_creation_status.ok()) {
+        LOG(INFO) << "Table creation status message: " <<
+        table_creation_status.message().ToString();
+      }
 
-    if (table_creation_status.message().ToString().find("Table already exists") ==
-        std::string::npos) {
-      CHECK_OK(table_creation_status);
+      if (table_creation_status.message().ToString().find("Table already exists") ==
+          std::string::npos) {
+        CHECK_OK(table_creation_status);
+      }
     }
 
     shared_ptr<YBTable> table;
@@ -603,31 +665,51 @@ int main(int argc, char* argv[]) {
 
     LOG(INFO) << "Starting load test";
     atomic_bool stop_flag(false);
-    MultiThreadedWriter writer(
-      FLAGS_load_test_num_rows,
-      FLAGS_load_test_num_writer_threads,
-      client.get(),
-      table.get(),
-      &stop_flag);
+    if (FLAGS_load_test_reads_only) {
+      SingleThreadedScanner scanner(table.get());
 
-    writer.Start();
+      scanner.countRows();
+    } else if (FLAGS_load_test_writes_only) {
+      SingleThreadedScanner scanner(table.get());
+      uint64 num = scanner.countRows();
 
-    MultiThreadedReader reader(
-      FLAGS_load_test_num_rows,
-      FLAGS_load_test_num_writer_threads,
-      client.get(),
-      table.get(),
-      writer.InsertionPoint(),
-      writer.InsertedKeys(),
-      writer.FailedKeys(),
-      &stop_flag);
-    reader.Start();
+      // Adds more keys starting from next index after scanned index
+      MultiThreadedWriter writer(
+          FLAGS_load_test_num_rows + num + 1, num + 1,
+          FLAGS_load_test_num_writer_threads,
+          client.get(),
+          table.get(),
+          &stop_flag);
 
-    writer.WaitForCompletion();
+      writer.Start();
+      writer.WaitForCompletion();
+    } else {
+      MultiThreadedWriter writer(
+        FLAGS_load_test_num_rows, 0,
+        FLAGS_load_test_num_writer_threads,
+        client.get(),
+        table.get(),
+        &stop_flag);
 
-    // The reader will not stop on its own, so we stop it as soon as the writer stops.
-    reader.Stop();
-    reader.WaitForCompletion();
+      writer.Start();
+      MultiThreadedReader reader(
+        FLAGS_load_test_num_rows,
+        FLAGS_load_test_num_writer_threads,
+        client.get(),
+        table.get(),
+        writer.InsertionPoint(),
+        writer.InsertedKeys(),
+        writer.FailedKeys(),
+        &stop_flag);
+
+      reader.Start();
+
+      writer.WaitForCompletion();
+
+      // The reader will not stop on its own, so we stop it as soon as the writer stops.
+      reader.Stop();
+      reader.WaitForCompletion();
+    }
 
     LOG(INFO) << "Test completed (iteration: " << i + 1 << " out of " <<
       FLAGS_load_test_num_iter << ")";
